ore: initial commit of completed Qt project
This commit adds the full project, which was previously developed but
not tracked in version control.

- All functionality is in place, including cloud request handling and UI
  setup.
- Replaced sensitive endpoint URLs and API keys with empty strings for
security. You may want to go through the codebase and replace the links
and keys with your own custom code. I didn't have time to add
environment variables.
- Added `.gitignore` entries for cache files and config files.

Note:
- Endpoint links and API keys have been replaced with empty strings
to avoid exposing sensitive information.